added redux to Resume.jsx

+++ src/pages/Resume/Resume.jsx
@@ -15,7 +15,7 @@ var samplesectors = [
     fileType: "txt",
     division: "Water",
     imageLoc: "blah/blah",
-    description: "I'm the best so I don't need to have any experience",
+    description: "I'm the best so I don't need to have any experience"
   },
   {
     sectorID: 2,
@@ -24,54 +24,64 @@ var samplesectors = [
     fileType: "txt",
     division: "Air",
     imageLoc: null,
-    description: "I'm the best so I don't need to have any projects",
-  },
-];
+    description: "I'm the best so I don't need to have any projects"
+  }
+]
+
 
 var sampleuser = {
   applicationUserId: 1,
   firstName: "Michael",
   lastName: "Chung",
   emailAddress: "mc@ae.com",
-};
+}
 
-// var curruser = {};
+var curruser = {};
 
 var resume = [];
+
+var index = 3;
+
+const client = axios.create({
+  baseURL: "localhost:5000/api/"
+});
+
+var noAPI = false;
 
 class Resume extends Component {
   constructor(props) {
     super(props);
-    this.state = {
-      currsector: undefined,
-      sectors: resume,
-      // user: sampleuser
-      user: {},
-    };
+    this.state = { currsector: undefined, sectors: resume, user: sampleuser };
   }
 
   componentDidMount() {
-    const curruser = axios.get(`/api/user/${this.props.userID}/`);
-    this.setState({ user: curruser });
-    axios
-      .get(`/api/user/${this.props.userID}/resume`)
-      .then((res) => {
-        resume = res.data.sectors;
-        this.setState({ sectors: resume });
-      })
-      .catch((err) => console.log(err));
+    if (noAPI) {
+      resume = samplesectors
+      curruser = sampleuser
+      this.setState({ sectors: resume });
+    } else {
+      //curruser = sampleuser;
+      this.setState({ user: curruser })
+      const url = `/api/user/${this.props.userID}/resume`
+      axios.get(url)
+        .then((res) => {
+          resume = res.data.sectors;
+          this.setState({ sectors: resume })
+        })
+    }
   }
 
   handleUpdate() {
-    axios
-      .get(`/api/user/${this.props.userID}/resume`)
-      .then((res) => {
-        resume = res.data.sectors;
-        this.setState({ sectors: resume });
-      })
-      .catch((err) => {
-        console.log(err);
-      });
+    if (noAPI) {
+      this.setState({ sectors: resume });
+    } else {
+      const url = `/api/user/${this.props.userID}/resume`
+      axios.get(url)
+        .then((res) => {
+          resume = res.data.sectors;
+          this.setState({ sectors: resume });
+        })
+    }
   }
 
   componentDidUpdate() {
@@ -90,40 +100,46 @@ class Resume extends Component {
       division: sectordivision,
       imageLoc: imageloc,
       description: sectordescription,
-    };
+    }
 
-    axios
-      .post(`/api/user/${this.props.userID}/resume/sector`, newsector)
-      .then((res) => {
-        this.handleUpdate();
+    if (noAPI) {
+      resume.push(newsector);
+      this.setState({ sectors: resume });
+    } else {
+      const url = `/api/user/${this.props.userID}/resume/sector`
+      axios.post(url, newsector).then((res) => {
+        this.handleUpdate()
       })
-      .catch((err) => {
-        console.log(err);
-      });
+    }
     console.log("Added " + newsector.name);
   }
 
   deleteSector(sector) {
-    const url = `/api/user/${this.props.userID}/resume/sector/${sector.sectorID}`;
-    axios.delete(url).then((res) => {
-      this.handleUpdate();
-    });
+    if (noAPI) {
+      resume.splice(resume.findIndex((e) => e.sectorID === sector.sectorID), 1);
+      this.setState({ sectors: resume });
+    } else {
+      const url = `/api/user/${this.props.userID}/resume/sector/${sector.sectorID}`
+      axios.delete(url).then((res) => {
+        this.handleUpdate()
+      })
+    }
     console.log("Deleted " + sector.name);
   }
 
   saveSector(sectorid, sectorname, sectorcontent) {
-    var oldsector = resume[resume.findIndex((e) => e.sectorID === sectorid)];
+    var oldsector = resume[resume.findIndex(e => e.sectorID === sectorid)]
     oldsector.name = sectorname;
     oldsector.description = sectorcontent;
-    const url = `/api/sector/${oldsector.sectorID}`;
-    axios
-      .put(url, oldsector)
-      .then((res) => {
-        this.handleUpdate();
+    if (noAPI) {
+      resume.splice(resume.findIndex(e => e.sectorID === sectorid), 1, oldsector);
+      this.setState({ sectors: resume });
+    } else {
+      const url = `/api/sector/${oldsector.sectorID}`
+      axios.put(url, oldsector).then((res) => {
+        this.handleUpdate()
       })
-      .catch((err) => {
-        console.log(err);
-      });
+    }
     console.log("Saved" + sectorid);
   }
 
@@ -135,41 +151,17 @@ class Resume extends Component {
         </div>
         <div className="resume-page">
           <div className="resume-builder">
-            <ResumeBuilder
-              sectors={this.state.sectors}
-              addSector={(
-                sectorname,
-                sectordivision,
-                filetype,
-                imageloc,
-                sectordescription
-              ) => {
-                this.addSector(
-                  sectorname,
-                  sectordivision,
-                  filetype,
-                  imageloc,
-                  sectordescription
-                );
-              }}
-              deleteSector={(sector) => {
-                this.deleteSector(sector);
-              }}
-              selectSector={(sectorid) => {
-                this.selectSector(sectorid);
-              }}
-            ></ResumeBuilder>
+            <ResumeBuilder sectors={this.state.sectors}
+              addSector={(sectorname, sectordivision, filetype, imageloc, sectordescription) => { this.addSector(sectorname, sectordivision, filetype, imageloc, sectordescription) }}
+              deleteSector={(sector) => { this.deleteSector(sector) }}
+              selectSector={(sectorid) => { this.selectSector(sectorid) }}></ResumeBuilder>
           </div>
           <div className="sector-editor">
             <div>
-              {this.state.currsector && (
-                <SectorEditor
-                  sector={this.state.currsector}
-                  saveSector={(sectorid, sectorname, sectorcontent) => {
-                    this.saveSector(sectorid, sectorname, sectorcontent);
-                  }}
-                ></SectorEditor>
-              )}
+              {this.state.currsector &&
+                <SectorEditor sector={this.state.currsector}
+                  saveSector={(sectorid, sectorname, sectorcontent) => { this.saveSector(sectorid, sectorname, sectorcontent) }}></SectorEditor>
+              }
             </div>
           </div>
         </div>
